♻️ Remove `test_constructor_invalid` test
The `test_constructor_invalid` test in the `data/pseudo/test_pseudo.py` module was
testing that the `PseudoPotentialData` constructor would fail with no input arguments.
This behaviour was the expected behavior of the parent `SinglefileData` class, from
which `PseudoPotentialData` inherits its constructor. However, the `SinglefileData`
constructor now accepts no input arguments since the changes in:

https://github.com/aiidateam/aiida-core/commit/958bfd05cabe403cd0ccf2832478d397c177685e

Which have been released in v2.7.0 of `aiida-core`. Hence, the test is no longer
correct/necessary, and is removed here.